Upgrade to Elasticsearch 5.1 (#177)
* Upgrade to Elasticsearch 5.1

* Fix sudo line for Travis CI

* Change mappings for Elasticsearch 5

The string type has been replaced by text and keyword types

* Fix apiVersion values

* Replace searchType="count"

The "count" searchType has been removed in Elasticsearch 5. Instead, we
do a normal search with size set to 0.

* Replace filtered query with bool query

The filtered query has been removed in Elasticsearch 5

* Only highlight the fields that we will use

* Handle change from terms filter to terms query

In Elasticsearch 5, there is no longer a terms filter. Instead, the
terms query can act in a filter context to accomplish the same purpose.
The accepted options are slightly different, so we have to get rid of
"execution" here.

* Update to new Elasticsearch domain